recover DBManager.

+++ TeamA/SmartAlbum/SmartAlbum/DB/DBManager.swift
@@ -22,12 +22,18 @@ import MapKit
 
 class DBManager {
     
-    static func initData(assets: [PHAsset], completionHandler: @escaping (Bool) -> Void) {
+    static func initData(assets: [PHAsset], isFirst:Bool, completionHandler: @escaping (Bool) -> Void) {
         let realm = try! Realm()
         
         var items:[Picture] = []
         //        assets.count
-        for i in 0..<300 {
+        var count = 0
+        if assets.count > 300 || isFirst {
+            count = 300
+        } else {
+            count = assets.count
+        }
+        for i in 0..<count {
             let asset = assets[i]
             let pic = Picture(asset: asset)
             
@@ -58,14 +64,14 @@ class DBManager {
                 } else {
                     //                    print("error in coreML")
                 }
-                
-                
                 items.append(pic)
             }
         }
         
         try! realm.write {
-            realm.deleteAll()
+            if isFirst {
+                realm.deleteAll()
+            }
             realm.add(items)
             print("add complete")
             //
@@ -74,136 +80,59 @@ class DBManager {
         }
     }
     
-    static func initData(assets: PHFetchResult<PHAsset>, completionHandler: @escaping (Bool) -> Void) {
+    static func initData(assets: PHFetchResult<PHAsset>, isFirst:Bool, completionHandler: @escaping (Bool) -> Void) {
         let assets2 = assets.objects(at: IndexSet(0..<assets.count))
-        initData(assets: assets2) { _ in
+        initData(assets: assets2, isFirst:isFirst) { _ in
             completionHandler(true)
         }
     }
     
-    static func initData(completionHandler: @escaping (Bool) -> Void) {
+    static func initData(isFirst:Bool, completionHandler: @escaping (Bool) -> Void) {
         // PHAsset 을 DB에 저장
         
         let fetchOptions = PHFetchOptions()
         fetchOptions.sortDescriptors = [NSSortDescriptor(key: "creationDate", ascending: false)]
         let assets:PHFetchResult<PHAsset> = PHAsset.fetchAssets(with: fetchOptions)
         
-        initData(assets: assets) { _ in
+        initData(assets: assets, isFirst:isFirst) { _ in
             completionHandler(true)
         }
        
     }
     
-    static func addLocation() {
-//        let realm = try! Realm()
-//
-//        let items = realm.objects(Picture.self).filter("location != nil")
-//        print("items.count:\(items.count)")
-//
-//        var i = 0
-//        for item in items {
-//            let id = item.value(forKey: "id") as! String
-//            let asset = PHAsset.fetchAssets(withLocalIdentifiers: [id], options: nil).firstObject
-//
-//            if let loc = asset?.location {
-//                LocationServices.getCity(location: loc) { city, error in
-//                    if error == nil {
-//                        try! realm.write {
-//                            i += 1
-//                            print("city \(i)")
-//                            item.city = city!
-//                        }
-//                    } else {
-//                        print("ERROR!")
-//                    }
-//                }
-//            }
-//        }
-        
-//        LocationServices.getCity{ _ in
-//            print("add location finished")
-//        }
-        
-    }
-    
-//    static func addLocation(completionHandler: @escaping (Bool) -> Void){
-//        let realm = try! Realm()
-//        let items = realm.objects(Picture.self).filter("location != nil")
-//        print("count:\(items.count)")
-//
-//        try! realm.write {
-//            for item in items {
-//                let loc = CLLocation(latitude: (item.location?.latitude)!, longitude: (item.location?.longtitude)!)
-//                LocationServices.getCity(location: loc){ (city, error) in
-//                    print("city:\(city!)")
-//                    item.location?.city = city!
-//                }
-//            }
-//        }
-//        print("add location complete")
-//        completionHandler(true)
-//    }
-    
     static func updateData(completionHandler: @escaping (Bool) -> Void) {
         // 사진의 modificationDate 값을 기준으로, 가장 최근에 keyword를 입힌 시점보다 최근에 수정된 이미지 파일은 keyword 업데이트
         // 첫 실행이 아닐 경우 무조건 updateData
+
 //        let realm = try! Realm()
+        
         let fetchOptions = PHFetchOptions()
         fetchOptions.sortDescriptors = [NSSortDescriptor(key: "modificationDate", ascending: false)]
-        let fetch:PHFetchResult<PHAsset> = PHAsset.fetchAssets(with: fetchOptions)
+        let fetchArr:PHFetchResult<PHAsset> = PHAsset.fetchAssets(with: fetchOptions)
+        let updateDate = UserDefaults.standard.value(forKey: "updateDate") as? Date
         
-        print("updateDate:\(UserDefaults.standard.value(forKey: "updateDate"))")
-        
-        var fetchedArray:[PHAsset] = []
-        for i in 0..<fetch.count {
-            let item = fetch.object(at: i)
-            print(item.modificationDate)
-            if item.modificationDate! < UserDefaults.standard.value(forKey: "updateDate") as! Date {
-//                fetchedArray = Array(fetch)[0..<i]
-                fetchedArray = fetch.objects(at: IndexSet(0..<i-1))
+        var newFetchArr:[PHAsset] = []
+        for i in 0..<fetchArr.count {
+            let fetch = fetchArr.object(at: i)
+            
+            if fetch.modificationDate! < updateDate! {
+                if i == 0 {
+                    break
+                }
+                newFetchArr = fetchArr.objects(at: IndexSet(0..<i-1))
                 break
             }
-            
-//            try! realm.write {
-//                let pic = Picture(asset: item)
-//                realm.add(pic)
-//            }
         }
         
-        initData(assets: fetchedArray) { _ in
-            print("\(fetchedArray.count) of items updated.")
+        initData(assets: newFetchArr, isFirst: false) { _ in
+            print("\(newFetchArr.count) of items updated.")
             UserDefaults.standard.set(Date(), forKey: "updateDate")
             completionHandler(true)
         }
-//        initData(assets: PHFetchResult<PHAsset>(fetchedArray), completionHandler: <#T##(Bool) -> Void#>)
-        
         // 삭제된 사진은 DB에서 지워야함.? -> 일단 보류
         
         
     }
-    
-//    static func getKeywords() -> [String] {
-//        //키워드 값들 목록 반환
-//        let realm = try! Realm()
-//        let keywords:[String] = Array(Set(realm.objects(Picture.self).value(forKey: "keyword") as! [String]))
-//        return keywords
-//    }
-    
-//    static func getLastAsset(_ keyword: String?) -> PHFetchResult<PHAsset> {
-//        let realm = try! Realm()
-//        let fetchOptions = PHFetchOptions()
-//        fetchOptions.sortDescriptors = [NSSortDescriptor(key: "creationDate", ascending: true)]
-//
-//        var identifier:String = ""
-//        if keyword == nil || keyword == "" {
-//            // 분류 안된 것만 리턴
-//            identifier = Array(realm.objects(Picture.self).filter("flag == 0")).last?.value(forKey: "id") as! String
-//        } else {
-//            identifier = Array(realm.objects(Picture.self).filter("keyword == %@", keyword!)).last?.value(forKey: "id") as! String
-//        }
-//        let fds:PHFetchResult<PHAsset> = PHAsset.fetchAssets(withLocalIdentifiers: [identifier], options: fetchOptions)
-//        return fds
-//    }
     
     static func getAssets(_ keyword: String?) -> PHFetchResult<PHAsset> {
         //키워드에 해당하는 PHAsset 불러옴
@@ -268,7 +197,7 @@ class DBManager {
         // 시 단위 그루핑
         let realm = try! Realm()
         
-        addLocation()
+//        addLocation()
         var citys:[String] = []
         var groupAssets:[PHFetchResult<PHAsset>] = []
         
@@ -365,7 +294,7 @@ class DBManager {
         for item in items {
             // 월단위 데이터 저장
             let itemDate = item.createDate //현재 아이템의 시간
-//            print("itemDate:\(item.createDate)")
+            print("itemDate:\(item.createDate)")
             
             if startDate.isInSameMonth(date: itemDate!) {
                 temp.append(item.value(forKey: "id") as! String)
@@ -374,11 +303,9 @@ class DBManager {
                 if temp.count != 0 {
                     groupAssets.append(getAssets(temp))
                     temp = []
-//                print("month itemDate:\(item.createDate)")
-//                print("month:\(startDate.getMonthString())")
                     groupKey.append(startDate.getMonthString())
                 }
-            
+                
                 startDate = startDate.getPrevMonth()
                 if startDate.isInSameMonth(date: itemDate!) {
                     temp.append(item.value(forKey: "id") as! String)
